+++ tests/mod.rs
@@ -16,6 +16,9 @@ use libcnb_test::BuildConfig;
 use std::env;
 use std::path::Path;
 
+// TODO: Switch to builder:26 once Ubuntu 26.04 is GA and also Python and package manager
+// versions have changed, so that we have a suitable historic buildpack version that both
+// supports the new stack, and has enough version changes for the cache invalidation tests.
 const DEFAULT_BUILDER: &str = "heroku/builder:24";
 
 fn default_build_config(fixture_path: impl AsRef<Path>) -> BuildConfig {
@@ -26,7 +29,9 @@ fn default_build_config(fixture_path: impl AsRef<Path>) -> BuildConfig {
     // to allow configuring the target arch independently of the builder name (eg via env var).
     let target_triple = match builder.as_str() {
         // Compile the buildpack for ARM64 iff the builder supports multi-arch and the host is ARM64.
-        "heroku/builder:24" if cfg!(target_arch = "aarch64") => "aarch64-unknown-linux-musl",
+        "heroku/builder:24" | "heroku/builder:26" if cfg!(target_arch = "aarch64") => {
+            "aarch64-unknown-linux-musl"
+        }
         _ => "x86_64-unknown-linux-musl",
     };
     config.target_triple(target_triple);